[refactor] User 객체 패키지 이동에 따른 import 수정 (#4)
