+++ src/main/java/konkuk/chacall/domain/chat/domain/ChatMessage.java
@@ -2,7 +2,7 @@ package konkuk.chacall.domain.chat.domain;
 
 import jakarta.persistence.*;
 import konkuk.chacall.domain.chat.domain.value.MessageContentType;
-import konkuk.chacall.domain.user.User;
+import konkuk.chacall.domain.user.domain.User;
 import konkuk.chacall.global.common.domain.BaseEntity;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
