+++ src/main/java/konkuk/chacall/domain/foodtruck/domain/FoodTruck.java
@@ -2,7 +2,7 @@ package konkuk.chacall.domain.foodtruck.domain;
 
 import jakarta.persistence.*;
 import konkuk.chacall.domain.foodtruck.domain.value.*;
-import konkuk.chacall.domain.user.User;
+import konkuk.chacall.domain.user.domain.User;
 import konkuk.chacall.global.common.converter.MenuCategoryListConverter;
 import konkuk.chacall.global.common.converter.PhotoUrlListConverter;
 import konkuk.chacall.global.common.domain.BaseEntity;
